Fix isentropic vortex ref state
