+++ test/Numerics/DGMethods/Euler/fvm_isentropicvortex.jl
@@ -25,6 +25,8 @@ const param_set = EarthParameterSet()
 
 using MPI, Logging, StaticArrays, LinearAlgebra, Printf, Dates, Test
 
+include("isentropicvortex_setup.jl")
+
 if !@isdefined integration_testing
     const integration_testing = parse(
         Bool,
@@ -151,10 +153,8 @@ function test_run(
         polynomialorder = (polynomialorder, 0),
     )
 
-    problem = AtmosProblem(
-        boundaryconditions = (),
-        init_state_prognostic = isentropicvortex_initialcondition!,
-    )
+    problem =
+        AtmosProblem(boundaryconditions = (), init_state_prognostic = setup)
 
     model = AtmosModel{FT}(
         AtmosLESConfigType,
@@ -186,7 +186,7 @@ function test_run(
     nsteps = ceil(Int, timeend / dt)
     dt = timeend / nsteps
 
-    Q = init_ode_state(dgfvm, FT(0), setup)
+    Q = init_ode_state(dgfvm, FT(0))
     lsrk = LSRK54CarpenterKennedy(dgfvm, Q; dt = dt, t0 = 0)
 
     eng0 = norm(Q)
@@ -253,66 +253,6 @@ function test_run(
     norm(Q - Qe) / norm(Qe) = %.16e
     """ level engf engf / eng0 engf - eng0 errf errf / engfe
     errf
-end
-
-Base.@kwdef struct IsentropicVortexSetup{FT}
-    p∞::FT = 10^5
-    T∞::FT = 300
-    ρ∞::FT = air_density(param_set, FT(T∞), FT(p∞))
-    translation_speed::FT = 150
-    translation_angle::FT = pi / 4
-    vortex_speed::FT = 50
-    vortex_radius::FT = 1 // 200
-    domain_halflength::FT = 1 // 20
-end
-
-function isentropicvortex_initialcondition!(
-    problem,
-    bl,
-    state,
-    aux,
-    localgeo,
-    t,
-    args...,
-)
-    setup = first(args)
-    FT = eltype(state)
-    x = MVector(localgeo.coord)
-
-    ρ∞ = setup.ρ∞
-    p∞ = setup.p∞
-    T∞ = setup.T∞
-    translation_speed = setup.translation_speed
-    α = setup.translation_angle
-    vortex_speed = setup.vortex_speed
-    R = setup.vortex_radius
-    L = setup.domain_halflength
-
-    u∞ = SVector(translation_speed * cos(α), translation_speed * sin(α), 0)
-
-    x .-= u∞ * t
-    # Make the function periodic
-    x .-= floor.((x .+ L) / 2L) * 2L
-
-    @inbounds begin
-        r = sqrt(x[1]^2 + x[2]^2)
-        δu_x = -vortex_speed * x[2] / R * exp(-(r / R)^2 / 2)
-        δu_y = vortex_speed * x[1] / R * exp(-(r / R)^2 / 2)
-    end
-    u = u∞ .+ SVector(δu_x, δu_y, 0)
-
-    _kappa_d::FT = kappa_d(param_set)
-    T = T∞ * (1 - _kappa_d * vortex_speed^2 / 2 * ρ∞ / p∞ * exp(-(r / R)^2))
-    # Adiabatic/isentropic relation
-    p = p∞ * (T / T∞)^(FT(1) / _kappa_d)
-    ts = PhaseDry_pT(bl.param_set, p, T)
-    ρ = air_density(ts)
-
-    e_pot = FT(0)
-    state.ρ = ρ
-    state.ρu = ρ * u
-    e_kin = u' * u / 2
-    state.ρe = ρ * total_energy(e_kin, e_pot, ts)
 end
 
 function do_output(
